+++ docs/content/docs/reference/widgets/Chart.md
@@ -16,7 +16,7 @@ Configure the widget using the builder methods named as follows:
 
 | Name | Type | Description |
 | --- | --- | --- |
-| **xAxis** | `PhpTui\Tui\Widget\Chart\Axis` | The X-Axis: bounds, style, labels etc. |
-| **yAxis** | `PhpTui\Tui\Widget\Chart\Axis` | The Y-Axis: bounds, style, labels etc. |
+| **xAxis** | `PhpTui\Tui\Extension\Core\Widget\Chart\Axis` | The X-Axis: bounds, style, labels etc. |
+| **yAxis** | `PhpTui\Tui\Extension\Core\Widget\Chart\Axis` | The Y-Axis: bounds, style, labels etc. |
 | **dataSets** | `DataSet[]` | The data sets. |
 | **style** | `PhpTui\Tui\Model\Style` | Style for the chart's area |
