+++ src/Extension/ImageMagick/ImageMagickExtension.php
@@ -3,6 +3,7 @@
 namespace PhpTui\Tui\Extension\ImageMagick;
 
 use PhpTui\Tui\Extension\ImageMagick\Shape\ImagePainter;
+use PhpTui\Tui\Extension\ImageMagick\Widget\ImageRenderer;
 use PhpTui\Tui\Model\DisplayExtension;
 
 final class ImageMagickExtension implements DisplayExtension
@@ -20,7 +21,9 @@ final class ImageMagickExtension implements DisplayExtension
 
     public function widgetRenderers(): array
     {
-        return [];
+        return [
+            new ImageRenderer($this->imageRegistry())
+        ];
     }
 
     private function imageRegistry(): ImageRegistry
